feat: finish transformer predictions
feat: find hyperparameters
feat: transform test data
feat: predict for test data
feat: add predictions
rearrange: rename custom_fix files
feat: add multiple trained transformer models
feat: add database with optimization studies (viz optuna)
