+++ predictions/transformer_test_final01.csv
@@ -1,0 +1,169 @@
+time,consumption
+2022-08-25 00:00:00+03:00,0.08640472456686935
+2022-08-25 01:00:00+03:00,0.15722676509833114
+2022-08-25 02:00:00+03:00,0.12315335500616664
+2022-08-25 03:00:00+03:00,0.1741466651110253
+2022-08-25 04:00:00+03:00,0.22542323139614426
+2022-08-25 05:00:00+03:00,0.13157663960741345
+2022-08-25 06:00:00+03:00,0.24922989678342805
+2022-08-25 07:00:00+03:00,0.3949085987876362
+2022-08-25 08:00:00+03:00,0.3462913471952602
+2022-08-25 09:00:00+03:00,0.5662520337436128
+2022-08-25 10:00:00+03:00,0.6937589344038357
+2022-08-25 11:00:00+03:00,0.47272238843333647
+2022-08-25 12:00:00+03:00,0.5918648053353028
+2022-08-25 13:00:00+03:00,0.7298014053108841
+2022-08-25 14:00:00+03:00,0.5867947928234294
+2022-08-25 15:00:00+03:00,1.0123180735203012
+2022-08-25 16:00:00+03:00,0.769722893690103
+2022-08-25 17:00:00+03:00,0.9055728650126881
+2022-08-25 18:00:00+03:00,0.34314658492308087
+2022-08-25 19:00:00+03:00,0.33384123468645965
+2022-08-25 20:00:00+03:00,0.3938966968876954
+2022-08-25 21:00:00+03:00,0.27850992692297677
+2022-08-25 22:00:00+03:00,0.3273704778448019
+2022-08-25 23:00:00+03:00,0.29167536199437494
+2022-08-26 00:00:00+03:00,0.07422962713136046
+2022-08-26 01:00:00+03:00,0.12848742446153774
+2022-08-26 02:00:00+03:00,0.11904311239585753
+2022-08-26 03:00:00+03:00,0.19543683621833122
+2022-08-26 04:00:00+03:00,0.237909254597968
+2022-08-26 05:00:00+03:00,0.2565881456742528
+2022-08-26 06:00:00+03:00,0.10693642349352249
+2022-08-26 07:00:00+03:00,0.16840488986364807
+2022-08-26 08:00:00+03:00,0.2159705991563467
+2022-08-26 09:00:00+03:00,0.21129529902199054
+2022-08-26 10:00:00+03:00,0.3599376236609795
+2022-08-26 11:00:00+03:00,0.22767353242654853
+2022-08-26 12:00:00+03:00,0.3262806899452852
+2022-08-26 13:00:00+03:00,0.3761884759283322
+2022-08-26 14:00:00+03:00,0.3519627678724158
+2022-08-26 15:00:00+03:00,0.32184412533503887
+2022-08-26 16:00:00+03:00,0.5824521848149569
+2022-08-26 17:00:00+03:00,0.5081527073929497
+2022-08-26 18:00:00+03:00,0.4600000494906613
+2022-08-26 19:00:00+03:00,0.41904162633152037
+2022-08-26 20:00:00+03:00,0.5259239758467565
+2022-08-26 21:00:00+03:00,0.36564795768939923
+2022-08-26 22:00:00+03:00,0.45651623893783305
+2022-08-26 23:00:00+03:00,0.3186827962210252
+2022-08-27 00:00:00+03:00,0.2175607788879905
+2022-08-27 01:00:00+03:00,0.27941063833458407
+2022-08-27 02:00:00+03:00,0.21340126817647698
+2022-08-27 03:00:00+03:00,0.2529836004240532
+2022-08-27 04:00:00+03:00,0.466221768416188
+2022-08-27 05:00:00+03:00,0.30529922634139217
+2022-08-27 06:00:00+03:00,0.29603856143308865
+2022-08-27 07:00:00+03:00,0.46379755338220013
+2022-08-27 08:00:00+03:00,0.39206838502412356
+2022-08-27 09:00:00+03:00,0.5079146011246312
+2022-08-27 10:00:00+03:00,0.7883988612110066
+2022-08-27 11:00:00+03:00,0.5947710362145278
+2022-08-27 12:00:00+03:00,0.7218670514757148
+2022-08-27 13:00:00+03:00,0.5633275566868625
+2022-08-27 14:00:00+03:00,0.8254601994677411
+2022-08-27 15:00:00+03:00,0.7741751751862729
+2022-08-27 16:00:00+03:00,0.8902156456114358
+2022-08-27 17:00:00+03:00,0.5488311457747655
+2022-08-27 18:00:00+03:00,0.7919621789003966
+2022-08-27 19:00:00+03:00,0.6726447585201744
+2022-08-27 20:00:00+03:00,0.8017360980756021
+2022-08-27 21:00:00+03:00,0.5596248833896751
+2022-08-27 22:00:00+03:00,0.6891756957937609
+2022-08-27 23:00:00+03:00,0.45153367006063155
+2022-08-28 00:00:00+03:00,0.13982607429514265
+2022-08-28 01:00:00+03:00,0.2075644604142098
+2022-08-28 02:00:00+03:00,0.18490786384829902
+2022-08-28 03:00:00+03:00,0.2880477039161537
+2022-08-28 04:00:00+03:00,0.36773631112024624
+2022-08-28 05:00:00+03:00,0.28786314773970595
+2022-08-28 06:00:00+03:00,0.32863141567336324
+2022-08-28 07:00:00+03:00,0.3974147833306497
+2022-08-28 08:00:00+03:00,0.3640686172321832
+2022-08-28 09:00:00+03:00,0.4874786970170025
+2022-08-28 10:00:00+03:00,0.7404680007731307
+2022-08-28 11:00:00+03:00,0.5594112171472999
+2022-08-28 12:00:00+03:00,0.6773407356994093
+2022-08-28 13:00:00+03:00,0.6487038082713995
+2022-08-28 14:00:00+03:00,0.8634855367441111
+2022-08-28 15:00:00+03:00,0.8430868552899446
+2022-08-28 16:00:00+03:00,1.108221090917721
+2022-08-28 17:00:00+03:00,0.804550572080879
+2022-08-28 18:00:00+03:00,0.8876676032688483
+2022-08-28 19:00:00+03:00,0.7404201134958035
+2022-08-28 20:00:00+03:00,0.6926356948789782
+2022-08-28 21:00:00+03:00,0.7506262088055943
+2022-08-28 22:00:00+03:00,0.8344241818902279
+2022-08-28 23:00:00+03:00,0.5890013860387397
+2022-08-29 00:00:00+03:00,0.16776663605237987
+2022-08-29 01:00:00+03:00,0.23642736433056077
+2022-08-29 02:00:00+03:00,0.21148852352673597
+2022-08-29 03:00:00+03:00,0.25719947792039133
+2022-08-29 04:00:00+03:00,0.5471239841382817
+2022-08-29 05:00:00+03:00,0.3776244509679974
+2022-08-29 06:00:00+03:00,0.44565457876090386
+2022-08-29 07:00:00+03:00,0.40689160297528537
+2022-08-29 08:00:00+03:00,0.4367544901813819
+2022-08-29 09:00:00+03:00,0.4397199244602976
+2022-08-29 10:00:00+03:00,0.7398744259716341
+2022-08-29 11:00:00+03:00,0.49251312977149964
+2022-08-29 12:00:00+03:00,0.8722473572826849
+2022-08-29 13:00:00+03:00,0.9952157702484149
+2022-08-29 14:00:00+03:00,0.8947890992521361
+2022-08-29 15:00:00+03:00,0.9175647138632922
+2022-08-29 16:00:00+03:00,1.0354536683284763
+2022-08-29 17:00:00+03:00,0.8128770753260103
+2022-08-29 18:00:00+03:00,0.7667519024248701
+2022-08-29 19:00:00+03:00,0.7015275397396196
+2022-08-29 20:00:00+03:00,0.553662079103303
+2022-08-29 21:00:00+03:00,0.6813358652815195
+2022-08-29 22:00:00+03:00,0.7499864653196884
+2022-08-29 23:00:00+03:00,0.49612695583791927
+2022-08-30 00:00:00+03:00,0.19691072991651348
+2022-08-30 01:00:00+03:00,0.25378404167338303
+2022-08-30 02:00:00+03:00,0.2029108063220507
+2022-08-30 03:00:00+03:00,0.29052283763611275
+2022-08-30 04:00:00+03:00,0.4041068034213
+2022-08-30 05:00:00+03:00,0.3079423665112503
+2022-08-30 06:00:00+03:00,0.5776118897516378
+2022-08-30 07:00:00+03:00,0.5811767514769581
+2022-08-30 08:00:00+03:00,0.5410883440461198
+2022-08-30 09:00:00+03:00,0.8208970025573913
+2022-08-30 10:00:00+03:00,0.8662706077451272
+2022-08-30 11:00:00+03:00,0.6415506600111491
+2022-08-30 12:00:00+03:00,0.6819009894493765
+2022-08-30 13:00:00+03:00,0.7177600603473259
+2022-08-30 14:00:00+03:00,0.6985644200405308
+2022-08-30 15:00:00+03:00,0.7662251396060844
+2022-08-30 16:00:00+03:00,0.9360769352055062
+2022-08-30 17:00:00+03:00,0.9436478502166069
+2022-08-30 18:00:00+03:00,0.7519125760880279
+2022-08-30 19:00:00+03:00,0.5492503351422455
+2022-08-30 20:00:00+03:00,0.5241498487600382
+2022-08-30 21:00:00+03:00,0.4652631140970606
+2022-08-30 22:00:00+03:00,0.5502571134124072
+2022-08-30 23:00:00+03:00,0.40673285957383665
+2022-08-31 00:00:00+03:00,0.30331713705398444
+2022-08-31 01:00:00+03:00,0.21926734983441903
+2022-08-31 02:00:00+03:00,0.2592457341348556
+2022-08-31 03:00:00+03:00,0.39140596450616466
+2022-08-31 04:00:00+03:00,0.4389715762687327
+2022-08-31 05:00:00+03:00,0.4670390939565299
+2022-08-31 06:00:00+03:00,0.5413896576966137
+2022-08-31 07:00:00+03:00,0.4885556722059724
+2022-08-31 08:00:00+03:00,0.5248212820890733
+2022-08-31 09:00:00+03:00,0.626375886950839
+2022-08-31 10:00:00+03:00,0.708676514111561
+2022-08-31 11:00:00+03:00,0.5706044836081391
+2022-08-31 12:00:00+03:00,0.37096719330144023
+2022-08-31 13:00:00+03:00,0.5555306608526337
+2022-08-31 14:00:00+03:00,0.48626109115989835
+2022-08-31 15:00:00+03:00,0.594932941036408
+2022-08-31 16:00:00+03:00,0.7582604202049582
+2022-08-31 17:00:00+03:00,0.646782206673433
+2022-08-31 18:00:00+03:00,0.6096828284031269
+2022-08-31 19:00:00+03:00,0.5128071969135505
+2022-08-31 20:00:00+03:00,0.6229258252052293
+2022-08-31 21:00:00+03:00,0.5599311963993222
+2022-08-31 22:00:00+03:00,0.4731083901056002
+2022-08-31 23:00:00+03:00,0.3728660002063376
